Moving validation code to a separate (sub) package
