+++ src/test/kotlin/cryptoservice/service/CryptoValidationTest.kt
@@ -1,5 +1,6 @@
 package cryptoservice.service
 
+import cryptoservice.service.validation.CryptoValidation
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
